+++ bcbio/ngsalign/postalign.py
@@ -25,19 +25,18 @@ def tobam_cl(data, out_file, is_paired=False):
     - If unpaired, use biobambam's bammarkduplicates
     """
     do_dedup = _check_dedup(data)
-    with tx_tmpdir(data) as tmpdir:
-        with file_transaction(data, out_file) as tx_out_file:
-            if not do_dedup:
-                yield (_sam_to_sortbam_cl(data, tmpdir, tx_out_file), tx_out_file)
-            elif is_paired:
-                sr_file = "%s-sr.bam" % os.path.splitext(out_file)[0]
-                disc_file = "%s-disc.bam" % os.path.splitext(out_file)[0]
-                with file_transaction(data, sr_file) as tx_sr_file:
-                    with file_transaction(data, disc_file) as tx_disc_file:
-                        yield (samblaster_dedup_sort(data, tmpdir, tx_out_file, tx_sr_file, tx_disc_file),
-                               tx_out_file)
-            else:
-                yield (_biobambam_dedup_sort(data, tmpdir, tx_out_file), tx_out_file)
+    with file_transaction(data, out_file) as tx_out_file:
+        if not do_dedup:
+            yield (_sam_to_sortbam_cl(data, tx_out_file), tx_out_file)
+        elif is_paired:
+            sr_file = "%s-sr.bam" % os.path.splitext(out_file)[0]
+            disc_file = "%s-disc.bam" % os.path.splitext(out_file)[0]
+            with file_transaction(data, sr_file) as tx_sr_file:
+                with file_transaction(data, disc_file) as tx_disc_file:
+                    yield (samblaster_dedup_sort(data, tx_out_file, tx_sr_file, tx_disc_file),
+                           tx_out_file)
+        else:
+            yield (_biobambam_dedup_sort(data, tx_out_file), tx_out_file)
 
 def _get_cores_memory(data, downscale=2):
     """Retrieve cores and memory, using samtools as baseline.
@@ -50,28 +49,24 @@ def _get_cores_memory(data, downscale=2):
                                          downscale, "decrease").upper()
     return num_cores, max_mem
 
-def _sam_to_sortbam_cl(data, tmpdir, tx_out_file):
-    """Convert to sorted BAM output with sambamba.
+def _sam_to_sortbam_cl(data, tx_out_file):
+    """Convert to sorted BAM output.
     """
     samtools = config_utils.get_program("samtools", data["config"])
-    sambamba = config_utils.get_program("sambamba", data["config"])
     cores, mem = _get_cores_memory(data, downscale=2)
-    return ("{samtools} view -b -S -u - | "
-            "{sambamba} sort -t {cores} -m {mem} "
-            "--tmpdir {tmpdir} -o {tx_out_file} /dev/stdin".format(**locals()))
+    tmp_file = "%s-sorttmp" % utils.splitext_plus(tx_out_file)[0]
+    return ("{samtools} sort -@ {cores} -m {mem} "
+            "-T {tmp_file} -o {tx_out_file} /dev/stdin".format(**locals()))
 
-def samblaster_dedup_sort(data, tmpdir, tx_out_file, tx_sr_file, tx_disc_file):
+def samblaster_dedup_sort(data, tx_out_file, tx_sr_file, tx_disc_file):
     """Deduplicate and sort with samblaster, produces split read and discordant pair files.
     """
-    sambamba = config_utils.get_program("sambamba", data["config"])
     samblaster = config_utils.get_program("samblaster", data["config"])
     samtools = config_utils.get_program("samtools", data["config"])
     cores, mem = _get_cores_memory(data, downscale=2)
-    for dname in ["spl", "disc", "full"]:
-        utils.safe_makedir(os.path.join(tmpdir, dname))
-    tobam_cmd = ("{samtools} view -S -u /dev/stdin | "
-                 "{sambamba} sort -t {cores} -m {mem} --tmpdir {tmpdir}/{dext} "
-                 "-o {out_file} /dev/stdin")
+    tmp_file = "%s-sorttmp" % utils.splitext_plus(tx_out_file)[0]
+    tobam_cmd = ("{samtools} sort -@ {cores} -m {mem} "
+                 "-T {tmp_file}-{dext} -o {out_file} /dev/stdin")
     splitter_cmd = tobam_cmd.format(out_file=tx_sr_file, dext="spl", **locals())
     discordant_cmd = tobam_cmd.format(out_file=tx_disc_file, dext="disc", **locals())
     dedup_cmd = tobam_cmd.format(out_file=tx_out_file, dext="full", **locals())
@@ -79,20 +74,17 @@ def samblaster_dedup_sort(data, tmpdir, tx_out_file, tx_sr_file, tx_disc_file):
            "| {dedup_cmd}")
     return cmd.format(**locals())
 
-def _biobambam_dedup_sort(data, tmpdir, tx_out_file):
+def _biobambam_dedup_sort(data, tx_out_file):
     """Perform streaming deduplication and sorting with biobambam's bammarkduplicates2.
     """
     samtools = config_utils.get_program("samtools", data["config"])
-    sambamba = config_utils.get_program("sambamba", data["config"])
     bammarkduplicates = config_utils.get_program("bammarkduplicates", data["config"])
-    base_tmp = os.path.join(tmpdir, os.path.splitext(os.path.basename(tx_out_file))[0])
     cores, mem = _get_cores_memory(data, downscale=2)
-    sort2_tmpdir = utils.safe_makedir(os.path.join(tmpdir, "sort2"))
-    return ("{samtools} view -b -S -u - |"
-            "{samtools} sort -n -o -@ {cores} -m {mem} - {base_tmp}-sort | "
-            "{bammarkduplicates} tmpfile={base_tmp}-markdup "
+    tmp_file = "%s-sorttmp" % utils.splitext_plus(tx_out_file)[0]
+    return ("{samtools} sort -n -@ {cores} -m {mem} -O bam -T {tmp_file}-namesort - | "
+            "{bammarkduplicates} tmpfile={tmp_file}-markdup "
             "markthreads={cores} level=0 | "
-            "{sambamba} sort -t {cores} -m {mem} --tmpdir {sort2_tmpdir} "
+            "{samtools} sort -@ {cores} -m {mem} -T {tmp_file}-finalsort "
             "-o {tx_out_file} /dev/stdin").format(**locals())
 
 def _check_dedup(data):