+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 - Support snpEff 4.1, including re-download of snpEff databases on demand if out
   of sync with older versions.
+- Pass ploidy to GATK HaplotypeCaller, supporting multiple ploidies and correct
+  calling of X/Y/MT chromosomes. Requires GATK 3.3.
 
 ## 0.8.5 (11 January 2015)