+++ HISTORY.md
@@ -9,6 +9,12 @@
   GEMINI, since it only handles bi-allelic inputs. Thanks to Pär Larsson.
 - Pass ploidy to GATK HaplotypeCaller, supporting multiple ploidies and correct
   calling of X/Y/MT chromosomes. Requires GATK 3.3.
+- Remove extra 'none' sample when calling tumor-only samples using
+  MuTect. Harmonizes headers with other tumor-only callers and enables
+  tumor-only ensemble calling. Thanks to Miika Ahdesmaki.
+- Perform variant prioritization as part of tumor-only calling, using population
+  based frequencies like 1000 genomes and ExAC and presence in known disease
+  causing databases like COSMIC and Clinvar.
 - Switch to samtools sort from sambamba sort during alignment streaming. Saves
   steps in processing and conversions on single sample no deduplication inputs.
 - On AWS, download inputs for S3 instead of streaming into fastq preparation to